init: open source caddy plugin (#1)
Open source the initial version of the Caddy plugin which is essentially
what we have already shared with the community in https://github.com/scionassociation/scion-caddy.

In follow up pull requests, we will update to the latest version of
https://github.com/scionproto/scion.

We might also want to consider using the https://github.com/Anapaya/quic-go
fork to support low MTU paths in the future.