network: do not leak goroutine on Close() (#5)
Fix a go routine leak caused by the
blockedListener. Caddy starts a new http.Server
for each time it loads the configuration. Behind
the scene, it shares the same listener for all the
servers.

Thus, we need the underlying connection to stay
alive and share it between multiple blocked
listeners. Unfortunately, the blockedListener did
not close until the underlying connection was
closed -> all http servers were blocked. We now
close the blockedListeners immedieatly, when we
receive the Close call, but still keep the
underlying connection alive until there is no
listener requiring it anymore.